+++ CHANGELOG.md
@@ -85,3 +85,6 @@ combination of some elementary symmetric polynomials
 
 * new function `schurCombination`, to get a symmetric polynomial as a linear 
 combination of some Schur polynomials
+
+* new functions `kostkaNumbersWithGivenLambda` and 
+`symbolicKostkaNumbersWithGivenLambda`, to get the Kostka numbers with parameter